Add Logo versions with different font
I have the feeling that the text (i.e. 'GUL') of the original logo
itself is a little hard to read on the original logos, so here is
a variant with a bigger and maybe easier to read fontface.

Also add the GIMP files used to create there, to make it easy to
change details.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>