meson: Adapt build specifications to work with meson 0.29
Ubuntu 16.04 comes with meson 0.29. If the package is to be built with
jenkins on the usual built machines and a more recent meson version
would be needed, this would require additional work on that machines.

These differences exist (except from fixed bugs), concerning libgul:
- meson does not automatically create build directory
- no 'to_int()' function (used for API version)
- libraries have to be individually specified

Maybe there is a difference for the doxygen creation, but this is not
finished yet, so it's not tested with 0.29.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>