meson: Add built target to create doxygen docs
It seems to be impossible to prevent meson 0.29 to put the doxygen
in the default targets list. Even if 'build_always : false' is set
in the custom_target() the target will be build anyhow.

Fixed with meson 0.38 and build_by_default parameter.

meson 0.29 does not allow to capture doxygen's output.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>