time_util: Add poll-free interruptible sleep
It is based on a std::condition_variable. Handling is about as simple as
the std::atomic version.

Also put a file to try this into the tests/ directory. This is not a
test as usual, but intended to be used with some online compiler to try
this alternative implementation of sleep.
(Of course it would have to be removed before the branch is pulled.)

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>