time_util: Mitigate spurious wakes in sleep2()
Unfortunately there is no way I can see to distinguish spurious wakes
from notification calls. So we need an additional place to store if the
sleep2 has been cancelled or not.

To mitigate spurious wakes normally one uses the return value of
wait_for() and wait_until() which is of type std::cv_status. But again
spurious wakes and notify_all() calls look identical.

As requested sleep2() returns if the whole time has been waited (or if
the sleep has been aborted).

In total the implementation is not not as clean and simple anymore. And
then again, I don't know when spurious wakes can occur at all. We never
saw them. Probably they occur in more complicated setups with a lot of
threads.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>