meson: Add options to enable tests/doxygen
Building the tests and doxygeneration is now turned off by default.
If you want to use tests or doxygen, this needs to be specified when
meson configures the individual build.

New options:
 test (boolean)
 doxy (boolean)

For example
  meson build -Dtests=true
generates a build directory where tests are enabled.

The options can also be changed after a build directory had been
created. Just call
  meson configure -Dtests=true

in the particular build directory. Note that old versions of meson have
this as separate executable and the call would be
  mesonconf -Dtests=true

Hint:
To get an overview of all possible options call in the build dir
  meson configure
(Old versions do not have the overview output.)

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>