meson: Circumvent meson 0.29 problem
Meson 0.29 can not just copy files to the build directory, it can just
configure them. Before this commit we split up... Meson prior 0.47 did
an empty configuration call to copy the files.

Starting with meson 0.47 we got an option to just copy files. And an
additional warning if we try to configure files but do nothing (i.e.
just copy them).

To avoid the warning with modern mesons but still keep functional with
old mesons there was a conditional branch in the meson.build, that
checked meson's version.

But unfortunately meson 0.29 is so old, that it can not even report it's
own version number.

So now we revert to 'always configure' the files, but to avoid the
warning of modern versions the configuration is not empty anymore, but
we replace a string with itself. Sigh.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>