Replace the polling implementation of sleep() by an extended version of Fini's
This commit makes the old sleep2() the new sleep(). There are four overloads:

bool sleep(double)
bool sleep(duration)
bool sleep(duration, SleepInterrupt &)
bool sleep(double, SleepInterrupt &)

I have taken the liberty of renaming SleepData to SleepInterrupt and adding
an assignment operator and a constructor from bool so that it can be used
almost like the previous atomic_bool version (see the example in the docs).
We should decide if we want to keep the member functions interrupt() and reset().