meson: Show all relevant build info in one place
Show the API and git versions as well as the 'correctly' set --prefix in
one place when the build directory is created by meson.

This commit belongs to

  1bdebc4 Add a way to query the library version

but somehow it got lost while rebasing/merging/force-pushing :-(

API and git version are shown aligned, so that - if we have git tags to
identify the API version - one can easily check that both are equal.

We could even issue a (meson) warning there, if the versions do not
agree.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>