meson: Check git version and API versions semantically
We now support two different tags in the git repository:
 - Version tags (starting with 'v')
 - External/proprietary tags (for example Doocs version tags)

The version tags have to follow this form (following https://semver.org):
"vM.m.p" where M, m, and p are decimal numbers of Major, minor, and patch
version numbers respectively. For example "v3.12.8".

Then the git version (git tag starting with 'v') is used to extract the
API version from it (i.e. Major.minor) and that is compared to the API
version specified in the meson.build. Build fails if they do not match.
(Does not fail if git is unavailable, as git is no build time dep.)

The external version tag (e.g. Doocs version) is unused. We talked
about this and came to no conclusions yet. So we have it here, but
ignore it in fact (which I think is what we should do).

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>