hexdump: Redesign stream interface
After a discussion with Soeren it seemed to be better to just return
strings from hexdump() always.

Further added the possibility to construct the dump directly on a stream
rather than in memory (i.e. in a stringstream).

Syntax summary:
   auto x = hexdump(mystring, prompt)           // returns std::string
   auto y = hexdump(mycontainer, prompt)        // returns std::string
   auto y = hexdump(mydata, mylength, prompt)   // returns std::string

   mystream << hexdump_stream(mydata, mylength, prompt)

   hexdump_stream(myfilestream, mydata, mylength, prompt)

The hexdump_stream() variant does not (yet) take containers or strings
directly. The reason is that they can be trivially used with the given
interface, just a bit more typing involved.
Implementing is easy (in fact I have the file here), but the whole
header files becomes a bit hard to read.

So for now, I propose this limited set of functions, if this is approved
we can add the other overloads if we want to.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>