meson: Increase default warning level
Set to the meson warning level with the most compiler warnings enabled.
And add -Wshadow, because if this warns we have at least a design
problem.

Also default set prefix directory to the usual path.

Both changes make it easier to generate a default build system. Options
that are usually set are now set by default.

Old:
      meson --warnlevel 3 --prefix /usr -Dcpp_args=-Wshadow build
New:
      meson build

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>