meson: Add doxygen-pages install target
The documentation pages can be installed to usually
  /usr/local/share/libgul/html/

Destination directory can be customized with the meson datadir option.
For example:
meson --datadir=../web/tesla/doocs/doocs_libs/

Note that normally the prefix dir has to be observed (and it is
prepended to the datadir), thus the relative path in the example.

Note that the subdirectory within the datadir is set to the project name
(i.e. 'libgul') and not some alias of it (i.e. 'gul').

Installation is done with rsync. If no rsync is available the target
will not be created.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>