hexdump: Correct behavior when C string literals are to be dumped
The code now splits cases into C-string-literals and containers.
Before it split between can-be-converted-to-string_view and containers.

std::string and gul::string_view are containers in this sense and can be
handled by the general STL-container code.

This also adds the ability to dump C string literals to a stream.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>