hexdump: Lengthen lifetime of temporary objects to dump
This fixes the problem that we could not dump temporaries, because the
lifetime of their interators would have ended at the point
hexdump_stream() returned. (And thus before the forwarder object has
been evaluated by operator<<.)

Const refs to temporaries keep the temporary alive for the lifetime of
the ref.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>