hexdump: Fix iterator lifetime bug in gcc
It seems that there is a bug in gcc (used 7.3.0) that prevents extension
of lifetime for references to anonymous temporaries.

Did not dive too deeply into it, because we do not really need to extend
the temporaries of an original call in this way. 'This way' means calls
to hexdump with two iterators. But it is technically quite involved to
get two iterators out of one temporary, so we just ignore it.

Example

  hexdump(std::array<...>{..}.begin(), ???.end())

Hard to get the end() iterator to the second parameter.

These still is the bug lurking when a temporary container is to be
dumped. Added a commented out test case.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>