hexdump: Fix lifetime problem with container argument
The only problem we have with the lifetime of the arguments is when a
complete container is to be dumped. It is unlikely, but users can pass a
temporary container to hexdump_stream() and when the container is
actually traversed (i.e. after operator<<) it is already freed.

So if a container that is a temporary is passed as argument we need to
copy that container and retain it until we are finished with dumping.
This is done by a new overload that just takes temporary containers.

The HexdumpParameterForward{} needs to potentially hold that container,
so it is expanded by a appropriate member.

To be able to fill the HexdumpParameterForward{} the const of the
members is shifted from directly in the object to their instantiation
places.

Note: The declval used as 'enable_if' in the new hexdump_stream()
overload is not too nice (i.e. ugly), but that was the only way I could
come up to detect if the argument is a temporary or not.
Any ideas welcome.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>