meson: Change default prefix to the default default
The usual prefix is /usr/local. This is also the default of meson.

That has been overwritten with a new default /usr to attain easy debian
package building.

But it is rather uncommon that a project installs directly into /usr.
That is usually reserved for package managers, normal local installs go
into the (of course) /usr/local subdirectory.

So the default is now changed from /usr to /usr/local

A corresponding check for debian packagability has always been there.
What is also added now is a hint what to do if debian packaging is desired.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>