meson: Allow intermediate versions in debian/changelog
If for some reason HEAD is not the tagged commit, this will be
represented by a version number increase, for example:
  NAME_3_2_1
has 6 commits on top after the tag. Now the revision put into
the changelog will be
  3.2.1-6

Dirty git repositories are ignored and not further marked.
Usually one should not build a package from a dirty repo.
This is not enforced, though.

Of course all the git related information is available within the libgul
itself, together with the 'dirtyness'.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>