meson: Allow old meson versions
The generation of debian/changelog has been done in a way, that old
meson versions do incorrectly.
Old means that the error is still in meson 0.40.0

[why]
The configure placeholders '@something@' are parsed recursively in older
meson versions. That behavior had been dropped between 0.41 and 0.46.
Now the recursion only happenes if it is specified.

[how]
The changelog template is now generated directly into a specific file,
that is concretely named, by the changelogger script.

After that the processing/generation of the debin/changelog continues as
before; embedded placeholders are replaced etc.

This is anyhow the better way to do it. The former version was a hack to
accomodate 'modern' meson syntax even with older ones (which obviously
failed).