git: Do not ignore hidden files
I'm not really sure why all hidden files (i.e. name starting with a dot)
were on the ignore list. Or to put it better: I see no reason for this.

If we remove this we can better see if there are anywhere any (hidden)
files, in the checked out repository, that deviate from the original.

Only thing I can imagine are strange files from an IDE someone is using,
but that files are better in the user's own gitignore, i.e.

    git config --global core.excludesfile ~/.gitignore_global

because they need to be ignored in all git repos anyhow.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>