meson: Build all tests in one monolithic executable
This helps to keep the build times down, because there is just one
executable to build and not a multitude of executables.

The caveat is that the output after running the tests with ninja is now not
very detailed. To get details one has to instruct catch2 to generate a
detailed report and look into that manually.

It is impossible to pass arguments to the test executable by calling it
via ninja, so meson has to be used (or the monolithic test just called
directly).

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>