meson: Use external versioning also for .so
If external versioning (a.k.a. DOOCS version numbering) is used the
dynamic libraries need to be versioned with that same scheme to avoid
problems if an old package is removed while a newer one is already
installed (while the API version was the same).

If two different external versions with the same API version are
installed, i.e. both install the same libgul.so.3.2 file, and one
(presumably the older version) is removed, the library files it has
installed in removed, but that file is still needed for the newer
version.

So the dynamic libraries need the same external versioning scheme to
make parallel install and uninstall independent of each other.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>