meson: Add ability to create inter-version packets
[why]
Usually if a new package with with minor fixes is packaged the version
number's last part (patch level) would be increased.

But if the version number is fixed due to outer constrains and can not
be changed (like when using git-external-tags (i.e. DOOCSVERSION*)) we
need a different alternative to produce a package with different version
number.

[how]
A new switch enables evaluation of a more complete git state. Not only
is the version number taken from the git tags, but also the number of
commits in the repository since that tag are counted and appended to the
version number. So we end up with versions like:

"libgul-18-11-6-p5"

This means that the package has been built from a repository that is 5
commits ahead of the tagged commit.

[note]
Git usually gives also the hash of that commit (as shown in the GIT
VERSION indicator), to make sure that the commit is what one thinks it
is, which is not guaranteed with force pushes.
But we skip this for brevity and depend on well behaved developers.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>