meson: Fix version deduction on standard packets
[why]
When standard debian packets have been built the wrong version
information has been written to the changelog file. The external tag
version was used instead of the api version tag.

The code expects a set git_external_num only on builds with
"-D deb-vers-ext=true". Unfortunately that variable was always filled.

[how]
Only fill it if appropriate.

[note]
Don't know how long that error was already present. As this is just an
error in the documentation of the packets, it was probably there from
the beginning...

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>