meson: Unify packaging handling for standard and DOOCS packets
After reconsidering the available flags and what is really different
between normal debian conforming packets and DOOCS conforming packets
the whole process could be streamlined.

Now one just specifies how a version donating git tag should look like
(e.g. 'v2.3' or 'DOOCSBLAH_18_7_12') and the whole packaging preparation
is done in a uniform manner.

Still the real API versions are encoded in 'v2.3' style git tags, but
that is just checked once to be the same as the number given in the main
build definition file.

Background and initializer was the reuse of changelogger in other
projects, so it had to become more flexible. More flexible made it
simpler too.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>