deb: Add internal versioning to changelogger
[why]
If the script is used in multiple disjunct repositories (i.e. libgul and
libslx) one needs a way to identify the real script version independent
from the repository.
Only then one can know if backporting is needed or not, and which
repository has the newest version.

[how]
Add a comment stating a version number and change date.
This is very weak, but probably enough.

[note]
Unrelated: Add help option

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>