deb: Ensure build configuration is current
[why]
The complete configuration for debian packaging is done with meson at
build configuration time. When things (e.g. git tags, file changes)
differ in the actual building process this will get by unnoticed. The
result might be wrongly versioned packets.

[how]
If main components of the configuration are not current at the time the
debian/rule is run, the packetizing process will be aborted. The user
has then to initiate a reinitialisation.

This could not really be done automatically, because part of the
outdated files are already evaluated by the packeting process when we
finally can check it.
Having the reconfiguration automatically is extremely difficult when
separate tools like makeDdeb are to be used.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>