deb: Correct generated changelog for exact revisions
[why]
For in-between releases the topmost entry in the changelog is HEAD, and
as a commit the person mentioned in the logfile is the commiter.

But if the release is a exact hit for a version tag, the topmost entry
in the changelog is ref/version (which happenes to point to HEAD), and
here the person mentioned is the tagger instead.

The list of refs has to be handled appropriately, which is split in the
top ref (head) and the following refs (tail).

For some strange reason the parameter to 'tail' has been wrong, and
instead of giving all-but-the-first revision, it gave all revisions.

The result empty were ghost entries in the changelog when HEAD is a
tagged version.

[how]
Correctly use only the second and all following refs.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>