meson: Fix determination of git tagged API version
[why]
The API version encoded in the main meson.build specification is checked
against the git tags that describe the API version.
The tag is extracted from git with (not quite but):

    git describe --tags --match 'v*'

This returns the full description of the current repository, for example

    v0.3
    v0.4-2-gd723fcd
    v0.2-15-8sdh752s-dirty

To get to the API version the initial 'v' has to be removed and
everyting starting from and including a '-' if present.

The removal of '-...' was missing in the code.

Why is unclear for me. Something must have changed with the generation
of this stuff, but I can't blame it on a specific commit.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>