make: Adapt ARCH name generation to Darwin
[why]
There are two issues with the generation of the ARCH tag on Darwin:
- sed does not have GNU extensions, i.e. \L (lower case)
- uname does not have -o

[how]
Instead of relying on sed to generate a lower cased string we use good
old tr to do this for ASCII chars at least. As the ARCH name usually
does not contain any chars but, this is fine.

Instead of uname -o we use the less powerful uname -s. This skips some
extension names, but that are not critical. But we have a change of the
ARCH name on Linux. Examples:
  x86_64-linux        (previous was: x86_64-linux-gnu)
  x86_64-darwin

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>