join_split: Remove UB from test cases and try to prevent misuse
[why]
Empty regular expressions (i.e. std::regex{""}) are either undefined
behavior (UB) (for POSIX regexes) or can lead to indefinite loops (for
ECMAScript regexes).

This surfaced because the regex implementation on OSX and Linux are
different and to be honest the OSX implementation is more correct.

The relevant definition for the usual ECMAScript regexes is here:

Terms
Each Alternative is either empty or is a sequence of Terms (with no separators between the Terms)

    Alternative ::
        [empty]
        Alternative Term

Empty Alternative always matches and does not consume any input.

Taken from https://en.cppreference.com/w/cpp/regex/ecmascript

With the Linux/Ubuntu implementation an Empty Alternative matches
always, but does consume one character. This effectively prevents
endless matching. If we do not advance but matched nothing, we
will be in the exact same situation in the next iteration ... thus
looping indefinitely.

[how]
First the examples with empty regexes are removed from the test cases,
because they are not useful and the result may differ from platform to
platform.

Second the split code breaks off, if it detects an endless loop.

Third a new test case just tests if empty regexes end in an infinite
loop (thus the test fails with timeout), or somehow breaks off.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>