time_util: Remove unused object
[why]
We have nothing in time_util.cc anymore, and it has just been kept if
sometime we will add something in it again. But still the object is
packed in the libraries, with no symbols.

Darwin's ranlib complains (rightfully so) if something without symbols is
packed into a library.

[how]
Remove the object, and reinsert it if we ever need it again.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>