make: Simplify meson extra arguments
[why]
As both arguments (compile and link time extra arguments) are always
passed together, this can be simplified to one variable that holds all
extra arguments.
(When meson is invoked this is to prepare the build system and is thus
for compile and link stages.)

Also the use of '-Dname=xxx' and the secondary form '-D name=xxx' is not
used consistently throughout the Makefile.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>