Add withinOrders() and withinAbs()
Often one wants to compare two floating point numbers. But if for example
one was cast to a different (even higher resolution) floating point type
and back, the numbers might be different as far as the comparison
operator is concerned.

    double a{0.100};
    long double b{a};
    b += 1.0l;
    double c = b; // use assignment to allow narrowing conversion
    c -= 1.0;
    assert(a == c); // FAILS

Usually one considers a number equal, if it is equal enough to some
degree. Here we add 2 functions that allow this to be done:

   * Difference in (fractions of) orders of magnitude in the values
   * Difference is lower than an absolute number

A third possibility to specify a percentage of difference that is
allowed is not implemented, because that always poses the problem of
'The percentage of what? Number a or number b'.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>