num_util: Allow integer arguments for withinAbs()
[why]
There is no reason to limit the arguments of withinAbs() to just
floating point types.

[how]
Integer types need a slightly different formula than floating point
types. As we have no 'if constexpr' these variants are encoded in an
ordinary if clause. It will be removed by the optimizer.

[note]
Also correct date in comment.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>