num_util: Fix withinOrders() with values of different sign
[why]
If the two numbers to be compared had different sign the results were
wrong.

[how]
Handle the resulting r value accordingly.

Note that std::abs() will be extremely simplified by the optimizer,
so no need to do if()s. In fact code with if() is more complicated
and can not be as optimized. I'm astonished.

Also add some test cases for such things.

[note]
Add test cases for special floating point values like NaN and INF.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>