num_util: Fix withinAbs() for unsigned integers
[why]
The function used std::abs() to prevent negative diff values (that would make no
sense). But that is not defined for any unsigned integer type. It compiles just
because implicit type conversions, but that would be a narrowing conversion.

[how]
Because we are stuck with C++14 we can not use constexpr if to use abs()
only when the template is instantiated for a signed type.
Without constexpr if we can try to use a partially specialized template,
but this again is not possible for ordinary functions, so we stuff the
function into an otherwise empty class.

A very very roundabout way to say (C++17 code follows):

    template<typename num_t>
    bool withinAbs(num_t a, num_t b, num_t diff) noexcept {
        if constexpr (not std::is_unsigned<num_t>::value) {
            diff = abs(diff); // Negative diff does not make sense
        }
        if (a > b) {

    [...]

But still, with SFINAE and 14 lines of code later we achieve the same
effect with C++14 that took 1 line in C++17.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>