deb: Allow version tags to be lightweight tags
When creating the debian/changelog we determine the author of the
individual revision automatically from git (of course).

Usually we distinguish between the person who decided that one concrete
commit is our new revision - and the author of that commit.

Example:
  I authored the last commit.
  You decided the repository is good enough for packaging.
  You stick the revision tag to HEAD and build the package.

In the changelog YOU should be the author of the package. Of course not
me, who just did the last commit.

This is possible because tags and commits are distinct objects in git. A
tag just points to a certain commit. Tags have their own tag-authors,
tag-dates and so on.

This works well, except you use a lightweight tag.
Lightweight tags are not really tags. They function like branch names.
There is no tag-author or tag-date.

This commit implements a workaround to make the process possible with
lightweight tags. If a certain version is only a lightweight tag instead
of a proper ('annotated') tag, the name and date are taken from the commit
in question.

Original patch date: Fri, 30 Nov 2018 11:00:14 +0100

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>