deb: Add explanation in changelog to lightweight tags
[why]
If lightweight tags are used to identify the releases the information
who did the packet maintenance is not available. Normally the tagger id
is used as maintainer id.

The previous commit allows lightweight tags to identify releases.
Although it is not desired to use lightweight tags it is possible that
people do use them. It is only administrative and not enforced
technically to use only annotated tags.

If a lightweight tags sneaks in the repository the changelog generation
will break (prior to previous commit).

This is the reason, that the previous commit has been resurrected from
the archives and is now pull requested again.

But then we have maintainer ids which are sometimes the tagger,
sometimes just the topmost committer and it is known which is which.

[how]
To prevent problems with lightweight tags we technically allow/provide
for them to be somehow handled (previous commit).

A note is added in the changelog file that the maintainer is unknown
and the topmost committer id is used instead (this commit).