num_util: Add abs()
[why]
The std::abs() function does not handle unsigned parameter types very
well, if at all. In some cases they are implicitly converted to their
signed counterparts, but this is not guaranteed (and often a call to
std::abs() with an unsigned argument just fails).

For usage in template work one sometimes just want to make sure that a
number is never negative. This is what abs() is for. But if the template
happens to be instantiated for a unsigned value the code breaks (most
likely).

[how]
This gul::abs() forwards most calls just to std::abs() and the implicit
conversions etc are valid (see std::abs() documentation).

If the argument type happens to be is_unsigned<>, the function just
returns the argument itself as its original type, no narrowing, no
value range loss.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>