num_util: Simplify within_orders()' explanation
[why]
I found the explanation rather technical and not describing the aim of
the function in a readable manner.

The explanation for the behavior if one operand is zero not very
concrete.

[how]
I find it very hard to come up with an explanation that is both correct
and can be understood by people that just search for a function to
compare.
After some attempts I rewrote the first part together with my wife as
someone how is not very mathematical. I believe it is now much more easy
to grasp the concept behind it (even if it is not exactly correct). The
real description follows to allow more insights.

[note]
I wonder why it has been written with pow() and ratio stuff
and not the simple abs(a - b) < (max(a, b) / orders).

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>