Add SlidingBuffer<>{}
A simple data buffer with a (semi) fixed size to use as sliding window on a data stream

The SlidingBuffer augments a std::array with very few extra functions to make it possible
to use it as sliding window on some incoming data.

The SlidingBuffer shares some properties with ring buffers: It has a fixed maximum size
and incoming elements are easily appended. But there is no way to pop elements out of
the buffer. They will drop out automatically if the capacity is not sufficient to hold
all encountered incoming elements.

This buffer is not intended for producer-consumer problems. If the elements access is
atomic it is threadsafe though.

Differing from arrays this buffer will start empty (with a size of zero).
Once the buffer has grown to its designated size, it will never change in size again,
except when 'clear'ed.

A typical application would be to analyze an incoming stream of elements in a finite
impulse response filter.

The following lists the differences to the std::array interface. All interface
items not mentioned here access the underlying std::array interface.

Member types:
  value_type          Type of the elements
  BaseT               Type of the underlying atd::array (i.e. std::array<value_type, ..>)

Member functions:
  Element access:
    push_back         Insert an element into the buffer
    operator[]        Access element relative to most recent element in buffer, bounds corrected, read only
    at                Note: Not overloaded, gives access to underlying std:array
    front             Access first (next to be pushed out) element
    back              Acces last (most recently pushed in) element
  Iterators:
    end, cend         Returns iterator to end of used space in std::array
    rbegin, crbegin   Returns reverse iterator to beginning of used space in std::array
  Capacity:
    size              Returns number of used elements
    capacity          Returns maximum number of elements
    filled            Checks whether the buffer is completely filled
  Modifiers:
    clear             Empty the buffer

Non-member functions:
  operator<<          Dump the raw data of the buffer to an ostream

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>