doxygen: Correct namespace of overview pages
[why]
Doxygen knows in which namespace a description is written in. If you
want to reference a class the refering description needs to be in that
namespace.
So as strange as it seems, the descriptions of the modules need to be in
the correct namespace (i.e. gul), so that the references to classes are
generated correctly. This affects only classes, not functions.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>