doxygen: Ignore string_view warnings
[why]
The string_view.h has not really doxygen comments put in, so while the
abstract stuff is generated correctly we get a lot of warnings from
doxygen.

If we run doxygen we want to see meaningful documentation warnings, like
when some functions or members have been forgotten to document. This is
hard to see if the useful warnings are lost in the noise of string_view.

If there is no warning for the documentation we want to change, the
build should be silent.

[how]
Just drop all string_view warnings from the output. Unfortunately
doxygen, inflexible as always, can not configured to do this. So we just
grep the lines away.

This is a bit complicated because we need to filter stderr.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>