string_view: Simplify type aliases
[why]
Doxygen has a problem to see if defaulted template parameters are the
same as specified template parameters.

[how]
The type aliases for typical types (e.g. gul::string_view) were
specified in a way to also give the optional template parameter although
there was no need to give it - it is the same as the default value.

Technically the type aliases are correct and even very explicit, but
unnecessarily so. Now we let the default value kick in (which is the
same as the previously specified value).

This is even more robust in case the default should ever change ;)

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>