doxygen: Let doxywrap correctly forward exit status
[why]
The build target should fail if and only if:
 - the preparation to call doxygen failed
 - the doxygen call failed
The exit status of the error stream filter has to be ignored.

[how]
Not that the call isn't complicated enough with filtering the standard
error stream, we also need to preserve the exit status in a suitable
manner.

This is very verbose if we use a shell that does not support pipefail.
So grudgingly the execution shell of doxywrap is changed to bash.

[note]
Also simplify error stream regex a bit.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>