num_util: Add clamp()
[why]
It's quite common to clamp/clip/coerce/limit a value to a certain range
of possible values, given by the limits. Maybe one wants to make shure
an array index is between 0 and size()-1. Or the priority value of
something is between -20 and 20.

[how]
This is mostly the C++17 implementation of std::clamp().

[note]
Sometimes I want to know if the value hit the limits or was okay. In my
code I have an additional out parameter for this. But let's stick to the
standard lib here?

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>