num_util: Remove assert from clamp
[why]
We can only compare two instances of the objects for 'less'. If we want
to allow to have gul::clamp(v, low, high) with low == high there is no
way to check that low and high are valid.

[how]
Remove the assert.
This is also the way it is done in boost::clamp(), and the reasoning is
also taken from there.

[note]
Also fixed comment in tests.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>