SlidingBuffer: Make it resizable
[why]
In contrast what was expected we have seen some use cases now, where a
variable capacity of the sliding buffer would be desirable. This would
for example allow to change filter lengths at runtime.

[how]
The SlidingBuffer is split into two flavours (invisible to the user):
std::array and std::vector based. If the user knows the capacity
beforehand an array is used. Otherwise a vector is taken as underlying
container. Apart from the ability of the later to resize()/reserve() all
is identical.

When a SlidingBuffer is resized at runtime, the following happens:
* Shrinking: The oldes excess elements are just dropped
* Growing: The capacity changes, but the (used) size not. It will grow
  gradually when elements are pushed, as in the startup phase.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>