SlidingBuffer: Make underlying container's ctors available
[why]
We want to be able to construct the buffer with a certain size in case
of std::vector.

[how]
Make all inherited constructors available. We do not really need all,
and some might even make not so much sense (at least at the moment). But
they also do not hurt.

[note]
For example it does not make sense to use an initializer list on the
SlidingBuffer to pre-fill it, because just the underlying container will
be filled, but the buffer will be empty after construction as always.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>