statistics: Remove StatisticsElement<> from library
[why]
This is a rather special case when someone needs a data type with this
setup (i.e. having a struct with member 'val' that holds the data).

[how]
Move it to the tests as example there (needs documentation later on).

Also remove the standard accessor for non-fundamental data types, as we
do not have a 'standard' non-fundamental type now anymore.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>