statistics: Remove _it suffix from iterator based functions
[why]
When we stick to two-iterator ranges (of identical type) and don't
support iterator/sentinel ranges, we can sufficiently distinguish which
template the user needs to have instantiated.

[how]
Just drop the extra part of the function names.

[note]
Somehow std::accumulate() leaks into our test namespace, even despite we
did not pull the namespace in explicitly. I'm not sure where this comes
from, but it could be an STL bug ;)
This happens (at least) with clang 6.0.0 and GCC 7.3.0.

Hmm, seems to be a feature. Including <numeric> is enough, after that
all parts therein are visible also without std::, through
_GLIBCXX_VISIBILITY. This is new to me.

Has to be mentioned in the docu later, that gul::accumulate must be
used.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>