statistics: Move ElementAccessor() out of anon namespace
[why]
I have the feeling that ElementAccessor() should be in the Doxygen
documentation, because it is in all statistics function signatures
visible in the docs.

[how]
Doxygen can not pull single elements out of the anonymous namespace to
document them, while others continue to be hidden (there is only a
global switch).

The only way is to put this function in the normal (i.e. gul) namespace.
I'm not really happy with this either, so maybe we want to revoke this
commit. Or not. Shrug.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>